merge

+++ webserver/reply.cpp
@@ -269,9 +269,9 @@ reply reply::stock_reply(reply::status_type status)
 
 void reply::add_header(reply *rep, const std::string &name, const std::string &value, bool replace)
 {
-	int num = rep->headers.size();
+	size_t num = rep->headers.size();
 	if (replace) {
-		for (int h = 0; h < num; h++) {
+		for (size_t h = 0; h < num; h++) {
 			if (boost::iequals(rep->headers[h].name, name)) {
 				rep->headers[h].value = value;
 				return;
@@ -284,8 +284,8 @@ void reply::add_header(reply *rep, const std::string &name, const std::string &v
 }
 
 void reply::add_header_if_absent(reply *rep, const std::string &name, const std::string &value) {
-	int num = rep->headers.size();
-	for (int h = 0; h < num; h++) {
+	size_t num = rep->headers.size();
+	for (size_t h = 0; h < num; h++) {
 		if (boost::iequals(rep->headers[h].name, name)) {
 			// is present
 			return;
